fix: Correct component name
